requirements.txt
urllib3
aiohttp
requests
pyyaml